requirements.txt
agno==2.0.4
aiohappyeyeballs==2.6.1
aiohttp==3.12.15
aiosignal==1.4.0
annotated-types==0.7.0
anyio==4.10.0
async-timeout==5.0.1
attrs==25.3.0
cachetools==5.5.2
certifi==2025.8.3
charset-normalizer==3.4.3
click==8.2.1
ddgs==9.5.5
discord==2.3.2
discord.py==2.6.3
docstring_parser==0.17.0
exceptiongroup==1.3.0
fastapi==0.116.1
frozenlist==1.7.0
gitdb==4.0.12
GitPython==3.1.45
google-auth==2.40.3
google-genai==1.36.0
greenlet==3.2.4
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
idna==3.10
lxml==6.0.1
markdown-it-py==4.0.0
mdurl==0.1.2
multidict==6.6.4
packaging==25.0
primp==0.15.0
propcache==0.3.2
pyasn1==0.6.1
pyasn1_modules==0.4.2
pydantic==2.11.7
pydantic-settings==2.10.1
pydantic_core==2.33.2
Pygments==2.19.2
python-dotenv==1.1.1
python-multipart==0.0.20
PyYAML==6.0.2
requests==2.32.5
rich==14.1.0
rsa==4.9.1
shellingham==1.5.4
smmap==5.0.2
sniffio==1.3.1
SQLAlchemy==2.0.43
starlette==0.47.3
tenacity==9.1.2
typer==0.17.4
typing-inspection==0.4.1
typing_extensions==4.15.0
urllib3==2.5.0
uvicorn==0.35.0
websockets==15.0.1
yarl==1.20.1
